Add the inline Content-Disposition header if it is not found.
I discovered some PDFs were still downloading. This was happening as
some requests get returned without specifying any Content-Disposition
header.

The default action for the browser is to download these requests.

The current code was looking for the header and only modifying it if it was found.

This would cause the code to miss these type of requests.